Rename draggable_sheet_extent_scope
