+++ lib/src/navigation/navigation_routes.dart
@@ -2,7 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import '../draggable/draggable_sheet_extent.dart';
-import '../draggable/draggable_sheet_extent_scope.dart';
+import '../draggable/draggable_sheet_position_scope.dart';
 import '../draggable/sheet_draggable.dart';
 import '../foundation/sheet_gesture_tamperer.dart';
 import '../foundation/sheet_physics.dart';
